Create "GopherCon 2015" series
I wanted to create a new series for the gophercon posts this year.

Future posts should be in a subdirectory gophercon-2015/, but the first
announcement will have to stay where it is to avoid breaking inbound links.

There is also a problem that the javascript gubbins which rewrites series
slugs to proper nouns mis capitalises "GopherCon" to "Gophercon", maybe this
can be fixed with a hack.

Change-Id: I42c88fb681c82af0cbb9337b21de66798ae822e1