Create "GopherCon 2015" series
I wanted to create a new series for the gophercon posts this year.

Future posts should be in a subdirectory gophercon-2015/, but the first
announcement will have to stay where it is to avoid breaking inbound links.

I've included a hack which correctly capitalises the slug in the left nav
but cannot figure out how to do the same for the summary page that the
slug send you to.