advent-2016: add 'context' post
This post was originally posted in multiple parts and requests for
comments drawn from a couple of core authorities on the topic.  This
is a revision based on comments and feedback primarily from Peter
Bourgon.

For example, the feedback that storing 'common log objects' in context
might end up giving users the wrong idea, even though they are
log-scoped, was taken to heart: this code was already trimmed down
from my real code where I did that, so I could simplify the example to
just store the IDs and still serve the purpose of illustration.

Finally, the original plan to include a summary of three applications
of context in this piece with outlinks back to the full medium posts
were instead changed to include the introduction and the first piece,
based on feedback and engagement with the first "introduction" piece I
posted.