Support latest Hugo version
Sometime between 0.15 (the version running in production) and 0.31
(the current version), Hugo requires you to specify which theme you
are using, which in our case is "gaweb".

Not sure how best to test this, but without this patch hugo renders an
empty page, and with it, hugo renders the blog server.

Fixes #236.
Fixes #238.